[12.11-11:46] [DEV] fix compil error [NEXT] merge + norme

+++ sources/cleaner/cleaner_command/get_type_command_bonus.c
@@ -72,7 +72,7 @@ int			get_cmd(t_cmd *cmd)
 	{
 		ft_printf("cmd->av[%d] = %s\n", i, cmd->av[i]);//DEBUG
 		ft_printf("cmd->ac = %d\n", cmd->ac);//DEBUG`
-		if (ft_strlen(cmd->av[i]) == 0 && i + 1 != cmd->ac)
+		if (ft_strlen(cmd->av[i]) == 0 && i + 1 != (size_t)cmd->ac)
 		{
 			i++;
 			continue ;
